Added `tapValueExtend` and `tapValueReduce`
Added the `tapValueExtend` and `tapValueReduce` properties. The former
(`true` by default) allows to modify the selected range of values, by
tapping on the slider below or above the selected range. On the other
hand, `tapValueReduce` (`false` by default) allows to modify the
selected range of values, by tapping on the slider within the selected
range. These two behaviours only apply to tapping events, and are not
e.g., relevant for when the user drags the selected range.